Move draw control API to core.py, ensure old draw control is preserved on toolbar draw control change
